refactor to monolith model file
