--- brainscore_vision/models/resnet50_primary_visual_cortex/test.py
@@ -1,8 +1,0 @@
-import pytest
-import brainscore_vision
-
-
-@pytest.mark.travis_slow
-def test_has_identifier():
-    model = brainscore_vision.load_model('resnet50_tutorial')
-    assert model.identifier == 'resnet50_tutorial'
